requirements.txt
pandas
google-cloud-bigquery
google-cloud-storage
gcsfs
pyarrow